rps_solver_app/requirements.txt
kaggle_environments
fastapi
uvicorn[standard]
pettingzoo
vec_noise
sqlmodel
python-telegram-bot
python-dotenv